refactor: 코드 수정 * 게스트 id 쿼리로 받기, * 리뷰 검증 로직 수정 * Redis Config 추가
