+++ backend/src/test/java/com/team3/airdnd/config/QuerydslConfig.java
@@ -1,4 +1,4 @@
-package com.team3.airdnd;
+package com.team3.airdnd.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
